Changes to quiver plotting code
We change the code to now normalise all arrows, and use alpha between 0
and 1 for an in-plane component of magnetization between 0 and 1. We
also allow standalone quiver animation, as opposed to having to plot it
alongside a quiver plot. This comes at the expense of removing the
ability to simultaneously plot a colour plot and quiver, but I hope to
fix this in the future.